[SAC-227][SQL] Discard the logic of supporting table column entities (#228)
## What changes were proposed in this pull request?

It has been headache and troublesome to keep the logic to support table column but having flag to disable it. Even we have dirty workaround to remove column attribute from created entities, which incurred a bug I had to fix.

We don't have any chance to turn it on, and once we support column entities I think it should be supported as default and shouldn't be able to turn it off.

This patch proposes to discard the logic to supporting table column entities at all.

## How was this patch tested?

Modified UTs, also manually tested.

This closes #227

Change-Id: Id19ec82317ef6fd491a4db0760141e5657d28cfa