[SAC-256][BUILD] Exclude Hadoop dependencies from Beeju (#257)
## What changes were proposed in this pull request?

This patch proposes to exclude Hadoop dependencies from Beeju and leverages Hadoop dependencies from Spark. (Beeju is in "test" scope, which won't make problem in usage.)

The base ground of this decision is, if Hadoop version conflict occurs between Beeju and Spark, we should rely on Spark which cannot be replaceable.

## How was this patch tested?

Existing UTs.

Closes #256

Change-Id: I5e589c9319e85aa7b545df45c29d60a604beacbd